requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.4
Django==3.2.12
django-crispy-forms==1.13.0
django-multiselectfield==0.1.12
djangorestframework==3.13.1
filelock==3.4.1
identify==2.4.4
importlib-metadata==4.8.3
importlib-resources==5.2.3
ldap3==2.9.1
mysqlclient==2.1.0
nodeenv==1.6.0
Pillow==9.0.1
platformdirs==2.4.0
pre-commit==2.17.0
psycopg2==2.9.3
pyasn1==0.4.8
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
tzdata==2022.1
virtualenv==20.13.3
zipp==3.6.0
